fix: tested and fixed migration functions
