+++ packages/vendure-plugin-primary-collection/src/index.ts
@@ -1,5 +1,5 @@
 export * from './primary-collection-plugin';
-export * from './util/export-current-primary-collections';
-export * from './util/save-primary-collections';
+export * from './util/export-primary-collections';
+export * from './util/update-primary-collections';
 export * from './util/helpers';
 export * from './api/primary-collections-helper.service';
